Lab restructuring 2
